Rename: 파일 경로 수정
Redis 에 저장하는 JwtRefreshToken 클래스의 경로를 변경함

Related to: #5
